#8: Add Job -Update job -update JobService.java -update IConstant.java,IErrorMessage -Add JobCapabilitiesRepository.java, JobKeySkillsRepository.java, TempSkillsRepository.java -replace @PrimaryKeyJoinColumn annotation with @joinColumn #5 : Database -Update create script

+++ server/src/main/java/io/litmusblox/server/service/impl/JobService.java
@@ -5,22 +5,23 @@
 package io.litmusblox.server.service.impl;
 
 import io.litmusblox.server.Constant.IConstant;
-import io.litmusblox.server.model.Company;
-import io.litmusblox.server.model.Job;
-import io.litmusblox.server.model.User;
-import io.litmusblox.server.repository.CompanyRepository;
-import io.litmusblox.server.repository.JobRepository;
-import io.litmusblox.server.repository.JobScreeningQuestionsRepository;
-import io.litmusblox.server.repository.UserRepository;
+import io.litmusblox.server.model.*;
+import io.litmusblox.server.Constant.IErrorMessages;
+import io.litmusblox.server.repository.*;
 import io.litmusblox.server.service.IJobService;
 import io.litmusblox.server.service.JobResponseBean;
 import io.litmusblox.server.service.JobWorspaceBean;
 import io.litmusblox.server.service.JobWorspaceResponseBean;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
+
+import javax.validation.ValidationException;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -47,6 +48,15 @@ public class JobService implements IJobService {
     @Autowired
     JobScreeningQuestionsRepository jobScreeningQuestionsRepository;
 
+    @Autowired
+    TempSkillsRepository tempSkillsRepository;
+
+    @Autowired
+    JobKeySkillsRepository jobKeySkillsRepository;
+
+    @Autowired
+    JobCapabilitiesRepository jobCapabilitiesRepository;
+
     @Override
     public JobResponseBean addJob(Job job, String pageName) throws Exception {//add job with respective pageName
 
@@ -59,9 +69,12 @@ public class JobService implements IJobService {
             return addJobOverview(job,oldJob);
         }else if(pageName.equalsIgnoreCase(IConstant.SCREENING_QUESTIONS)){
             return addJobScreeningQuestions(job, oldJob);
-        }else {
+        }else if(pageName.equalsIgnoreCase(IConstant.SKILLS)){
+            return addJobKeySkills(job, oldJob);
+        } else {
             //throw an operation not supported exception
         }
+
         return null;
     }
 
@@ -125,7 +138,7 @@ public class JobService implements IJobService {
             job.setCreatedOn(new Date());
             job.setMlDataAvailable(false);
             //TODO: Remove the following piece of code and set the user & company as obtained from login
-            User u = userRepository.getOne(2L);
+            User u = userRepository.getOne(1L);
             job.setCreatedBy(u);
             Company c = companyRepository.getOne(1L);
             job.setCompanyId(c);
@@ -140,13 +153,85 @@ public class JobService implements IJobService {
 
     private JobResponseBean addJobScreeningQuestions(Job job, Job oldJob){ //method for add screening questions
 
-        if(null!=oldJob.getJobScreeningQuestionsList()){
+        if(job.getJobScreeningQuestionsList().size()>IConstant.SCREENING_QUESTIONS_LIST_MAX_SIZE){
+            throw new ValidationException(IErrorMessages.SCREENING_QUESTIONS_VALIDATION_MESSAGE+job.getId());
+        }
+        if(null!=oldJob.getJobScreeningQuestionsList() && oldJob.getJobScreeningQuestionsList().size()>0){
             jobScreeningQuestionsRepository.deleteAll(oldJob.getJobScreeningQuestionsList());//delete old job screening question list
         }
-        oldJob.getJobScreeningQuestionsList().addAll(job.getJobScreeningQuestionsList());//add new screening question list
+        //TODO:User is need to change
+        User u = userRepository.getOne(1L);
+
+        job.getJobScreeningQuestionsList().forEach(n->{n.setCreatedBy(u);n.setCreatedOn(new Date());n.setUpdatedOn(new Date());n.setUpdatedBy(u);});
+        jobScreeningQuestionsRepository.saveAll(job.getJobScreeningQuestionsList());
+        JobResponseBean jb=new JobResponseBean();
+        jb.setJobId(job.getId());
+        return jb;
+    }
+
+    private JobResponseBean addJobKeySkills(Job job, Job oldJob){ //update and add new key skill
+        if(null!=oldJob.getJobKeySkillsList() && oldJob.getJobKeySkillsList().size()>0){
+            throw new ValidationException("Job key skills "+ IErrorMessages.EMPTY_AND_NULL_MESSAGE + oldJob.getId());
+        }
+        List<TempSkills> tempSkillsList = new ArrayList<>();
+
+        // update ML_PROVIDED value from true to false
+        jobKeySkillsRepository.updateJobKeySkills(false, true,job.getId());
+
+        List<JobKeySkills> falseJobKeySkillslist = jobKeySkillsRepository.findByJobIdAndMlProvided(job.getId(), false);
+
+        //delete all key skills where MlProvided=false
+        jobKeySkillsRepository.deleteAll(falseJobKeySkillslist);
+
+        for (JobKeySkills jobKeySkills: job.getJobKeySkillsList()) {
+            if(null!=jobKeySkills.getId()){
+                jobKeySkills.setSelcted(true);
+            }else{
+                List<TempSkills> tempList = tempSkillsRepository.findAll();
+                if(tempList.contains(jobKeySkills.getSkillId().getSkillsMaster())){
+                    continue;
+                }
+                TempSkills tempSkills=new TempSkills();
+                tempSkills.setReviewed(false);
+                tempSkills.setSkillName(jobKeySkills.getSkillId().getSkillsMaster());
+                tempSkills = tempSkillsRepository.save(tempSkills);
+                tempSkillsList.add(tempSkills);
+            }
+        }
+
+        for (TempSkills tempSkills:tempSkillsList) {
+            JobKeySkills jobKeySkills=new JobKeySkills();
+            jobKeySkills.setMlProvided(false);
+            jobKeySkills.setSelcted(true);
+            jobKeySkills.setCreatedOn(new Date());
+            jobKeySkills.setSkillIdFromTemp(tempSkills);
+        }
+
+        oldJob.getJobKeySkillsList().addAll(job.getJobKeySkillsList());
         jobRepository.save(oldJob);
         JobResponseBean jb=new JobResponseBean();
         jb.setJobId(job.getId());
         return jb;
     }
+
+    private JobResponseBean addJobCapabilities(Job job){ //add job capabilities
+
+        if(null!=job.getJobCapabilityList() && job.getJobCapabilityList().size()>0){
+            throw new ValidationException("Job Capabilities "+ IErrorMessages.EMPTY_AND_NULL_MESSAGE + job.getId());
+        }
+        List<Long> capabilityList = new ArrayList<>();
+        job.getJobCapabilityList().forEach(n->capabilityList.add(n.getId()));
+
+        //update all capability list as unselected
+        jobCapabilitiesRepository.updateJobCapabilitiesForUnSelected(false, job.getId());
+
+        //update all capability list as selected
+        jobCapabilitiesRepository.updateJobCapabilitiesForSelected(true,job.getId(),capabilityList);
+
+        JobResponseBean jb=new JobResponseBean();
+        jb.setJobId(job.getId());
+        return jb;
+    }
+
+
 }
